+++ foo.mars
@@ -1,13 +1,6 @@
-macro print_str
-    stack_rev
-    @counter 0 def
-    @size stack_size def
-    while counter size < do
-        print_ascii
-        @counter counter 1 + def
-    end
+macro do_err
+    123
+    switch
 end
 
-"Hello, World!\n"
-switch Hello_World_
-print_str
+do_err
